refactor(sdk): split monolithic types file into domain-specific modules
Split 1,138-line src/types/index.ts into 9 focused modules:
- core.ts: execution modes, invocation I/O
- logger.ts: logger interface and contexts
- step.ts: step configuration and retry logic
- child-context.ts: child context configuration
- callback.ts: callback types and configuration
- invoke.ts: function invocation configuration
- wait-condition.ts: conditional wait types
- batch.ts: batch operations (map/parallel/concurrent)
- durable-context.ts: main DurableContext interface

Benefits:
- Improved navigation and maintainability
- Reduced merge conflicts
- Clearer dependencies
- All JSDoc documentation preserved
- 100% backward compatible

All 635 tests pass. No breaking changes.
