+++ packages/aws-durable-execution-sdk-js/src/types/child-context.ts
@@ -1,0 +1,21 @@
+import { Serdes } from "../utils/serdes/serdes";
+import { DurableContext } from "./durable-context";
+
+/**
+ * Configuration options for child context operations
+ */
+export interface ChildConfig<T> {
+  /** Serialization/deserialization configuration for child context data */
+  serdes?: Serdes<T>;
+  /** Sub-type identifier for categorizing child contexts */
+  subType?: string;
+  /** Function to generate summaries for large results (used internally by map/parallel) */
+  summaryGenerator?: (result: T) => string;
+}
+
+/**
+ * Function to be executed in a child context with isolated state
+ * @param context - DurableContext with isolated step counter and state tracking
+ * @returns Promise resolving to the child function result
+ */
+export type ChildFunc<T> = (context: DurableContext) => Promise<T>;
