+++ packages/aws-durable-execution-sdk-js/src/types/index.ts
@@ -1,1138 +1,9 @@
-import { Context } from "aws-lambda";
-import { TerminationManager } from "../termination-manager/termination-manager";
-import { ExecutionState } from "../storage/storage-provider";
-import { Serdes } from "../utils/serdes/serdes";
-
-import { ErrorObject, Operation } from "@aws-sdk/client-lambda";
-
-// Define BatchItemStatus enum
-export enum BatchItemStatus {
-  SUCCEEDED = "SUCCEEDED",
-  FAILED = "FAILED",
-  STARTED = "STARTED",
-}
-
-// Define DurableExecutionMode enum
-export enum DurableExecutionMode {
-  ExecutionMode = "ExecutionMode",
-  ReplayMode = "ReplayMode",
-  ReplaySucceededContext = "ReplaySucceededContext",
-}
-
-export interface LambdaHandler<T> {
-  (event: T, context: Context): Promise<DurableExecutionInvocationOutput>;
-}
-
-// TODO - prefer to import this entire input model from the SDK,
-// but it's not part of the frontend model so it doesn't get generated.
-export interface DurableExecutionInvocationInput {
-  DurableExecutionArn: string;
-  CheckpointToken: string;
-  InitialExecutionState: {
-    Operations: Operation[];
-    NextMarker: string;
-  };
-  LoggingMode?: string;
-
-  /**
-   * Flag to indicate if this execution is running against local runner.
-   */
-  LocalRunner?: boolean;
-}
-
-export enum InvocationStatus {
-  SUCCEEDED = "SUCCEEDED",
-  FAILED = "FAILED",
-  PENDING = "PENDING",
-}
-
-export enum OperationSubType {
-  STEP = "Step",
-  WAIT = "Wait",
-  CALLBACK = "Callback",
-  RUN_IN_CHILD_CONTEXT = "RunInChildContext",
-  MAP = "Map",
-  MAP_ITERATION = "MapIteration",
-  PARALLEL = "Parallel",
-  PARALLEL_BRANCH = "ParallelBranch",
-  WAIT_FOR_CALLBACK = "WaitForCallback",
-  WAIT_FOR_CONDITION = "WaitForCondition",
-  CHAINED_INVOKE = "ChainedInvoke",
-}
-
-interface DurableExecutionInvocationOutputFailed {
-  Status: InvocationStatus.FAILED;
-  Error: ErrorObject;
-}
-
-interface DurableExecutionInvocationOutputSucceeded {
-  Status: InvocationStatus.SUCCEEDED;
-  Result?: string;
-}
-
-interface DurableExecutionInvocationOutputPending {
-  Status: InvocationStatus.PENDING;
-}
-
-export type DurableExecutionInvocationOutput =
-  | DurableExecutionInvocationOutputSucceeded
-  | DurableExecutionInvocationOutputFailed
-  | DurableExecutionInvocationOutputPending;
-
-export interface DurableContext {
-  /**
-   * The underlying AWS Lambda context
-   */
-  lambdaContext: Context;
-  /**
-   * Logger instance for this context, enriched with execution context information
-   */
-  logger: Logger;
-  /**
-   * Executes a function as a durable step with automatic retry and state persistence
-   *
-   * @remarks
-   * **IMPORTANT**: `step()` is designed for single atomic operations and cannot be used to group
-   * multiple durable operations (like other steps, waits, or child contexts). The step function
-   * receives a simple `StepContext` (for logging only), not a full `DurableContext`.
-   *
-   * **To group multiple durable operations, use `runInChildContext()` instead:**
-   * ```typescript
-   * // ❌ WRONG: Cannot call durable operations inside step
-   * await context.step("process-order", async (ctx) => {
-   *   await context.wait(1000);              // ERROR: context not available
-   *   await context.step(async () => ...);   // ERROR: context not available
-   *   return result;
-   * });
-   *
-   * // ✅ CORRECT: Use runInChildContext to group operations
-   * await context.runInChildContext("process-order", async (childCtx) => {
-   *   await childCtx.wait(1000);
-   *   const step1 = await childCtx.step(async () => validateOrder(order));
-   *   const step2 = await childCtx.step(async () => chargePayment(step1));
-   *   return step2;
-   * });
-   * ```
-   *
-   * @param name - Step name for tracking and debugging
-   * @param fn - Function to execute as a durable step
-   * @param config - Optional configuration for retry strategy, semantics, and serialization
-   * @example
-   * ```typescript
-   * // ✅ Good: Single atomic operation
-   * const result = await context.step(
-   *   "fetch-user-data",
-   *   async (ctx) => {
-   *     return await fetchUserFromAPI(userId);
-   *   },
-   *   { retryStrategy: exponentialBackoff }
-   * );
-   * ```
-   */
-  step<T>(
-    name: string | undefined,
-    fn: StepFunc<T>,
-    config?: StepConfig<T>,
-  ): Promise<T>;
-
-  /**
-   * Executes a function as a durable step with automatic retry and state persistence
-   *
-   * @remarks
-   * **IMPORTANT**: `step()` cannot group multiple durable operations. Use `runInChildContext()`
-   * to group steps, waits, or other durable operations together. See the named overload for details.
-   *
-   * @param fn - Function to execute as a durable step
-   * @param config - Optional configuration for retry strategy, semantics, and serialization
-   * @example
-   * ```typescript
-   * const result = await context.step(
-   *   async (ctx) => {
-   *     return await fetchUserFromAPI(userId);
-   *   }
-   * );
-   * ```
-   */
-  step<T>(fn: StepFunc<T>, config?: StepConfig<T>): Promise<T>;
-
-  /**
-   * Invokes another durable function with the specified input
-   * @param name - Step name for tracking and debugging
-   * @param funcId - Function ID or ARN of the durable function to invoke
-   * @param input - Input data to pass to the invoked function
-   * @param config - Optional configuration for serialization
-   * @example
-   * ```typescript
-   * const result = await context.invoke(
-   *   "process-payment",
-   *   "arn:aws:lambda:us-east-1:123456789012:function:payment-processor",
-   *   { amount: 100, currency: "USD" }
-   * );
-   * ```
-   */
-  invoke<I, O>(
-    name: string,
-    funcId: string,
-    input: I,
-    config?: InvokeConfig<I, O>,
-  ): Promise<O>;
-
-  /**
-   * Invokes another durable function with the specified input
-   * @param funcId - Function ID or ARN of the durable function to invoke
-   * @param input - Input data to pass to the invoked function
-   * @param config - Optional configuration for serialization
-   * @example
-   * ```typescript
-   * const result = await context.invoke(
-   *   "payment-processor-function",
-   *   { amount: 100, currency: "USD" }
-   * );
-   * ```
-   */
-  invoke<I, O>(
-    funcId: string,
-    input: I,
-    config?: InvokeConfig<I, O>,
-  ): Promise<O>;
-
-  /**
-   * Runs a function in a child context with isolated state and execution tracking
-   * @param name - Step name for tracking and debugging
-   * @param fn - Function to execute in the child context
-   * @param config - Optional configuration for serialization and sub-typing
-   * @example
-   * ```typescript
-   * const result = await context.runInChildContext(
-   *   "process-batch",
-   *   async (childCtx) => {
-   *     // Child context has its own step counter and state
-   *     const step1 = await childCtx.step("validate", async () => validate(data));
-   *     const step2 = await childCtx.step("transform", async () => transform(step1));
-   *     return step2;
-   *   },
-   *   { subType: "batch-processor" }
-   * );
-   * ```
-   */
-  runInChildContext<T>(
-    name: string | undefined,
-    fn: ChildFunc<T>,
-    config?: ChildConfig<T>,
-  ): Promise<T>;
-
-  /**
-   * Runs a function in a child context with isolated state and execution tracking
-   * @param fn - Function to execute in the child context
-   * @param config - Optional configuration for serialization and sub-typing
-   * @example
-   * ```typescript
-   * const result = await context.runInChildContext(
-   *   async (childCtx) => {
-   *     return await childCtx.step(async () => processData(input));
-   *   }
-   * );
-   * ```
-   */
-  runInChildContext<T>(fn: ChildFunc<T>, config?: ChildConfig<T>): Promise<T>;
-
-  /**
-   * Pauses execution for the specified duration
-   * @param name - Step name for tracking and debugging
-   * @param millis - Duration to wait in milliseconds
-   * @example
-   * ```typescript
-   * // Wait 5 seconds before retrying
-   * await context.wait("retry-delay", 5000);
-   * ```
-   */
-  wait(name: string, millis: number): Promise<void>;
-
-  /**
-   * Pauses execution for the specified duration
-   * @param millis - Duration to wait in milliseconds
-   * @example
-   * ```typescript
-   * // Wait 30 seconds for rate limiting
-   * await context.wait(30000);
-   * ```
-   */
-  wait(millis: number): Promise<void>;
-
-  /**
-   * Waits for a condition to be met by periodically checking state
-   * @param name - Step name for tracking and debugging
-   * @param checkFunc - Function that checks the current state and returns updated state
-   * @param config - Configuration for initial state, wait strategy, and serialization
-   * @example
-   * ```typescript
-   * const finalState = await context.waitForCondition(
-   *   "wait-for-job-completion",
-   *   async (currentState, ctx) => {
-   *     const jobStatus = await checkJobStatus(currentState.jobId);
-   *     return { ...currentState, status: jobStatus };
-   *   },
-   *   {
-   *     initialState: { jobId: "job-123", status: "pending" },
-   *     waitStrategy: (state, attempt) => {
-   *       if (state.status === "completed") {
-   *         return { shouldContinue: false };
-   *       }
-   *       return { shouldContinue: true, delaySeconds: Math.min(attempt * 2, 60) };
-   *     }
-   *   }
-   * );
-   * ```
-   */
-  waitForCondition<T>(
-    name: string | undefined,
-    checkFunc: WaitForConditionCheckFunc<T>,
-    config: WaitForConditionConfig<T>,
-  ): Promise<T>;
-
-  /**
-   * Waits for a condition to be met by periodically checking state
-   * @param checkFunc - Function that checks the current state and returns updated state
-   * @param config - Configuration for initial state, wait strategy, and serialization
-   * @example
-   * ```typescript
-   * const result = await context.waitForCondition(
-   *   async (state, ctx) => {
-   *     const updated = await pollExternalAPI(state.requestId);
-   *     return updated;
-   *   },
-   *   {
-   *     initialState: { requestId: "req-456", ready: false },
-   *     waitStrategy: (state, attempt) =>
-   *       state.ready ? { shouldContinue: false } : { shouldContinue: true, delaySeconds: 10 }
-   *   }
-   * );
-   * ```
-   */
-  waitForCondition<T>(
-    checkFunc: WaitForConditionCheckFunc<T>,
-    config: WaitForConditionConfig<T>,
-  ): Promise<T>;
-
-  /**
-   * Creates a callback that external systems can complete
-   * @param name - Step name for tracking and debugging
-   * @param config - Optional configuration for timeout and serialization
-   * @returns Tuple of [promise that resolves when callback is submitted, callback ID]
-   * @example
-   * ```typescript
-   * const [callbackPromise, callbackId] = await context.createCallback(
-   *   "external-approval",
-   *   { timeout: 3600 } // 1 hour timeout
-   * );
-   *
-   * // Send callback ID to external system
-   * await sendApprovalRequest(callbackId, requestData);
-   *
-   * // Wait for external system to submit callback
-   * const approvalResult = await callbackPromise;
-   * ```
-   */
-  createCallback<T>(
-    name: string | undefined,
-    config?: CreateCallbackConfig<T>,
-  ): Promise<CreateCallbackResult<T>>;
-
-  /**
-   * Creates a callback that external systems can complete
-   * @param config - Optional configuration for timeout and serialization
-   * @returns Tuple of [promise that resolves when callback is submitted, callback ID]
-   * @example
-   * ```typescript
-   * const [promise, callbackId] = await context.createCallback({
-   *   timeout: 1800 // 30 minutes
-   * });
-   * await notifyExternalSystem(callbackId);
-   * const result = await promise;
-   * ```
-   */
-  createCallback<T>(
-    config?: CreateCallbackConfig<T>,
-  ): Promise<CreateCallbackResult<T>>;
-  /**
-   * Wait for an external system to complete a callback with the SendDurableExecutionCallbackSuccess or SendDurableExecutionCallbackFailure APIs.
-   * @param name - Step name for tracking and debugging
-   * @param submitter - Function that receives the callback ID and submits the callback
-   * @param config - Optional configuration for timeout and retry behavior
-   * @example
-   * ```typescript
-   * const result = await context.waitForCallback(
-   *   "wait-for-external-api",
-   *   async (callbackId, ctx) => {
-   *     // Submit callback ID to external system
-   *     await submitToExternalAPI(callbackId);
-   *   },
-   *   { timeout: 300 }
-   * );
-   * ```
-   */
-  waitForCallback<T>(
-    name: string | undefined,
-    submitter: WaitForCallbackSubmitterFunc,
-    config?: WaitForCallbackConfig<T>,
-  ): Promise<T>;
-
-  /**
-   * Wait for an external system to complete a callback with the SendDurableExecutionCallbackSuccess or SendDurableExecutionCallbackFailure APIs.
-   * @param submitter - Function that receives the callback ID and submits the callback
-   * @param config - Optional configuration for timeout and retry behavior
-   * @example
-   * ```typescript
-   * const result = await context.waitForCallback(
-   *   async (callbackId, ctx) => {
-   *     await submitToExternalAPI(callbackId);
-   *   }
-   * );
-   * ```
-   */
-  waitForCallback<T>(
-    submitter: WaitForCallbackSubmitterFunc,
-    config?: WaitForCallbackConfig<T>,
-  ): Promise<T>;
-  /**
-   * Maps over an array of items with a function, executing in parallel with optional concurrency control
-   * @param name - Step name for tracking and debugging
-   * @param items - Array of items to process
-   * @param mapFunc - Function to apply to each item (context, item, index, array) =\> Promise\<TOutput\>
-   * @param config - Optional configuration for concurrency, completion behavior, and item naming
-   * @example
-   * ```typescript
-   * const results = await context.map(
-   *   "process-users",
-   *   users,
-   *   async (ctx, user, index) => processUser(user),
-   *   {
-   *     maxConcurrency: 2,
-   *     itemNamer: (user, index) => `User-${user.id || index}`
-   *   }
-   * );
-   * ```
-   */
-  map<TInput, TOutput>(
-    name: string | undefined,
-    items: TInput[],
-    mapFunc: MapFunc<TInput, TOutput>,
-    config?: MapConfig<TInput, TOutput>,
-  ): Promise<BatchResult<TOutput>>;
-
-  /**
-   * Maps over an array of items with a function, executing in parallel with optional concurrency control
-   * @param items - Array of items to process
-   * @param mapFunc - Function to apply to each item (context, item, index, array) =\> Promise\<TOutput\>
-   * @param config - Optional configuration for concurrency and completion behavior
-   * @example
-   * ```typescript
-   * const results = await context.map(
-   *   [1, 2, 3],
-   *   async (ctx, item, index) => item * 2
-   * );
-   * ```
-   */
-  map<TInput, TOutput>(
-    items: TInput[],
-    mapFunc: MapFunc<TInput, TOutput>,
-    config?: MapConfig<TInput, TOutput>,
-  ): Promise<BatchResult<TOutput>>;
-  /**
-   * Executes multiple functions in parallel with optional concurrency control
-   * @param name - Step name for tracking and debugging
-   * @param branches - Array of functions or named branches to execute in parallel
-   * @param config - Optional configuration for concurrency and completion behavior
-   * @example
-   * ```typescript
-   * const results = await context.parallel(
-   *   "parallel-operations",
-   *   [
-   *     { name: "task1", func: async (ctx) => ctx.step(async () => "result1") },
-   *     async (ctx) => ctx.step(async () => "task2")
-   *   ],
-   *   { maxConcurrency: 2 }
-   * );
-   * ```
-   */
-  parallel<T>(
-    name: string | undefined,
-    branches: (ParallelFunc<T> | NamedParallelBranch<T>)[],
-    config?: ParallelConfig<T>,
-  ): Promise<BatchResult<T>>;
-
-  /**
-   * Executes multiple functions in parallel with optional concurrency control
-   * @param branches - Array of functions or named branches to execute in parallel
-   * @param config - Optional configuration for concurrency and completion behavior
-   * @example
-   * ```typescript
-   * const results = await context.parallel(
-   *   [
-   *     async (ctx) => ctx.step(async () => "task1"),
-   *     async (ctx) => ctx.step(async () => "task2")
-   *   ]
-   * );
-   * ```
-   */
-  parallel<T>(
-    branches: (ParallelFunc<T> | NamedParallelBranch<T>)[],
-    config?: ParallelConfig<T>,
-  ): Promise<BatchResult<T>>;
-  promise: {
-    /**
-     * Waits for all promises to resolve and returns an array of all results
-     *
-     * @remarks
-     * **IMPORTANT**: Promise combinators accept already-created promises that start executing immediately.
-     * They cannot control concurrency, implement completion policies, or provide durability.
-     *
-     * **Consider using `map()` or `parallel()` instead if you need:**
-     * - Concurrency control (limit simultaneous executions)
-     * - Completion policies (minSuccessful, toleratedFailureCount)
-     * - Durability (survive Lambda timeouts and resume from checkpoints)
-     * - Per-item retry strategies
-     * - Progress tracking for long-running operations
-     *
-     * **Use promise combinators only for:**
-     * - Fast, in-memory operations (less than a few seconds)
-     * - Operations that must all start immediately
-     * - Simple coordination of already-running promises
-     *
-     * @param name - Step name for tracking and debugging
-     * @param promises - Array of promises to wait for (already executing)
-     * @example
-     * ```typescript
-     * // ❌ All promises start immediately - no concurrency control
-     * const [user, posts, comments] = await context.promise.all(
-     *   "fetch-user-data",
-     *   [
-     *     fetchUser(userId),        // Already running
-     *     fetchUserPosts(userId),   // Already running
-     *     fetchUserComments(userId) // Already running
-     *   ]
-     * );
-     *
-     * // ✅ Better: Use map() for controlled, durable execution
-     * const results = await context.map(
-     *   [userId, userId, userId],
-     *   async (ctx, id, index) => {
-     *     if (index === 0) return fetchUser(id);
-     *     if (index === 1) return fetchUserPosts(id);
-     *     return fetchUserComments(id);
-     *   },
-     *   { maxConcurrency: 2 } // Control concurrency, survives timeouts
-     * );
-     * ```
-     */
-    all<T>(name: string | undefined, promises: Promise<T>[]): Promise<T[]>;
-
-    /**
-     * Waits for all promises to resolve and returns an array of all results
-     *
-     * @remarks
-     * **IMPORTANT**: Promises start executing immediately when created. Consider using `map()` or `parallel()`
-     * for concurrency control, durability, and completion policies. See the named overload for details.
-     *
-     * @param promises - Array of promises to wait for (already executing)
-     * @example
-     * ```typescript
-     * // ❌ Limited: No concurrency control or durability
-     * const results = await context.promise.all([
-     *   fetchUser(userId),
-     *   fetchUserPosts(userId)
-     * ]);
-     *
-     * // ✅ Better: Use parallel() for durable execution
-     * const results = await context.parallel([
-     *   async (ctx) => ctx.step(async () => fetchUser(userId)),
-     *   async (ctx) => ctx.step(async () => fetchUserPosts(userId))
-     * ]);
-     * ```
-     */
-    all<T>(promises: Promise<T>[]): Promise<T[]>;
-
-    /**
-     * Waits for all promises to settle (resolve or reject) and returns results with status
-     *
-     * @remarks
-     * **IMPORTANT**: Promise combinators accept already-created promises that start executing immediately.
-     * Consider using `map()` or `parallel()` with completion policies for better control over failure handling.
-     *
-     * @param name - Step name for tracking and debugging
-     * @param promises - Array of promises to wait for (already executing)
-     * @example
-     * ```typescript
-     * // ❌ All promises start immediately
-     * const results = await context.promise.allSettled(
-     *   "fetch-all-data",
-     *   [
-     *     fetchUser(userId),
-     *     fetchUserPosts(userId),
-     *     fetchUserComments(userId)
-     *   ]
-     * );
-     *
-     * // ✅ Better: Use map() with completion config
-     * const results = await context.map(
-     *   [userId, userId, userId],
-     *   async (ctx, id, index) => {
-     *     // Fetch different data based on index
-     *   },
-     *   {
-     *     completionConfig: {
-     *       minSuccessful: 2, // Stop early if 2 succeed
-     *       toleratedFailureCount: 1
-     *     }
-     *   }
-     * );
-     * ```
-     */
-    allSettled<T>(
-      name: string | undefined,
-      promises: Promise<T>[],
-    ): Promise<PromiseSettledResult<T>[]>;
-
-    /**
-     * Waits for all promises to settle (resolve or reject) and returns results with status
-     *
-     * @remarks
-     * **IMPORTANT**: Promises start executing immediately. Consider using `map()` or `parallel()`
-     * for better failure handling and durability. See the named overload for details.
-     *
-     * @param promises - Array of promises to wait for (already executing)
-     */
-    allSettled<T>(promises: Promise<T>[]): Promise<PromiseSettledResult<T>[]>;
-
-    /**
-     * Waits for the first promise to resolve successfully, ignoring rejections until all fail
-     *
-     * @remarks
-     * **IMPORTANT**: Promise combinators accept already-created promises that start executing immediately.
-     * All promises race simultaneously with no control over execution order or resource usage.
-     *
-     * **Consider using `parallel()` with completion policies instead** for controlled racing with durability.
-     *
-     * @param name - Step name for tracking and debugging
-     * @param promises - Array of promises to race (already executing)
-     * @example
-     * ```typescript
-     * // ❌ All sources queried immediately
-     * const userData = await context.promise.any(
-     *   "fetch-from-any-source",
-     *   [
-     *     fetchFromPrimaryDB(userId),
-     *     fetchFromSecondaryDB(userId),
-     *     fetchFromCache(userId)
-     *   ]
-     * );
-     *
-     * // ✅ Better: Use parallel() with early completion
-     * const result = await context.parallel(
-     *   [
-     *     async (ctx) => ctx.step(async () => fetchFromPrimaryDB(userId)),
-     *     async (ctx) => ctx.step(async () => fetchFromSecondaryDB(userId)),
-     *     async (ctx) => ctx.step(async () => fetchFromCache(userId))
-     *   ],
-     *   {
-     *     completionConfig: { minSuccessful: 1 } // Stop after first success
-     *   }
-     * );
-     * ```
-     */
-    any<T>(name: string | undefined, promises: Promise<T>[]): Promise<T>;
-
-    /**
-     * Waits for the first promise to resolve successfully, ignoring rejections until all fail
-     *
-     * @remarks
-     * **IMPORTANT**: Promises start executing immediately. Consider using `parallel()` with
-     * `minSuccessful: 1` for durable racing. See the named overload for details.
-     *
-     * @param promises - Array of promises to race (already executing)
-     */
-    any<T>(promises: Promise<T>[]): Promise<T>;
-
-    /**
-     * Returns the result of the first promise to settle (resolve or reject)
-     *
-     * @remarks
-     * **IMPORTANT**: Promise combinators accept already-created promises that start executing immediately.
-     * All promises race simultaneously with no control over execution.
-     *
-     * **Use promise.race() only for:**
-     * - Racing against timeouts or deadlines
-     * - Simple coordination of already-running operations
-     *
-     * **For durable racing with control, use `parallel()` with `minSuccessful: 1`**
-     *
-     * @param name - Step name for tracking and debugging
-     * @param promises - Array of promises to race (already executing)
-     * @example
-     * ```typescript
-     * // ✅ Good use case: Racing against timeout
-     * const result = await context.promise.race(
-     *   "fetch-with-timeout",
-     *   [
-     *     fetchFromAPI(userId),
-     *     new Promise((_, reject) =>
-     *       setTimeout(() => reject(new Error("Timeout")), 5000)
-     *     )
-     *   ]
-     * );
-     * ```
-     */
-    race<T>(name: string | undefined, promises: Promise<T>[]): Promise<T>;
-
-    /**
-     * Returns the result of the first promise to settle (resolve or reject)
-     *
-     * @remarks
-     * **IMPORTANT**: Promises start executing immediately. Use only for simple timeout patterns.
-     * See the named overload for details.
-     *
-     * @param promises - Array of promises to race (already executing)
-     */
-    race<T>(promises: Promise<T>[]): Promise<T>;
-  };
-
-  /**
-   * Executes items concurrently with fine-grained control over execution strategy
-   * @param name - Step name for tracking and debugging
-   * @param items - Array of items to execute concurrently
-   * @param executor - Function that processes each item
-   * @param config - Optional configuration for concurrency and completion behavior
-   * @example
-   * ```typescript
-   * const results = await context.executeConcurrently(
-   *   "process-files",
-   *   [
-   *     { id: "file1", path: "/path/to/file1.txt" },
-   *     { id: "file2", path: "/path/to/file2.txt" },
-   *     { id: "file3", path: "/path/to/file3.txt" }
-   *   ],
-   *   async (item, ctx) => {
-   *     return await ctx.step(`process-${item.id}`, async () => {
-   *       return await processFile(item.path);
-   *     });
-   *   },
-   *   {
-   *     maxConcurrency: 3,
-   *     completionConfig: {
-   *       minSuccessful: 2,
-   *       toleratedFailureCount: 1
-   *     }
-   *   }
-   * );
-   * ```
-   */
-  executeConcurrently<TItem, TResult>(
-    name: string | undefined,
-    items: ConcurrentExecutionItem<TItem>[],
-    executor: ConcurrentExecutor<TItem, TResult>,
-    config?: ConcurrencyConfig<TResult>,
-  ): Promise<BatchResult<TResult>>;
-
-  /**
-   * Executes items concurrently with fine-grained control over execution strategy
-   * @param items - Array of items to execute concurrently
-   * @param executor - Function that processes each item
-   * @param config - Optional configuration for concurrency and completion behavior
-   * @example
-   * ```typescript
-   * const results = await context.executeConcurrently(
-   *   [{ data: "item1" }, { data: "item2" }],
-   *   async (item, ctx) => {
-   *     return await processItem(item.data);
-   *   }
-   * );
-   * ```
-   */
-  executeConcurrently<TItem, TResult>(
-    items: ConcurrentExecutionItem<TItem>[],
-    executor: ConcurrentExecutor<TItem, TResult>,
-    config?: ConcurrencyConfig<TResult>,
-  ): Promise<BatchResult<TResult>>;
-
-  /**
-   * Sets a custom logger for this context
-   * @param logger - Custom logger implementation
-   * @example
-   * ```typescript
-   * const customLogger = {
-   *   log: (level, message, data, error) => console.log(`[${level}] ${message}`, data),
-   *   error: (message, error, data) => console.error(message, error, data),
-   *   warn: (message, data) => console.warn(message, data),
-   *   info: (message, data) => console.info(message, data),
-   *   debug: (message, data) => console.debug(message, data)
-   * };
-   * context.setCustomLogger(customLogger);
-   * ```
-   */
-  setCustomLogger(logger: Logger): void;
-}
-
-export interface ExecutionContext {
-  state: ExecutionState;
-  _stepData: Record<string, Operation>; // Private, use getStepData() instead
-  terminationManager: TerminationManager;
-  durableExecutionArn: string;
-  getStepData(stepId: string): Operation | undefined;
-}
-
-export interface RetryDecision {
-  shouldRetry: boolean;
-  delaySeconds?: number;
-}
-
-export enum StepSemantics {
-  AtMostOncePerRetry = "AT_MOST_ONCE_PER_RETRY",
-  AtLeastOncePerRetry = "AT_LEAST_ONCE_PER_RETRY",
-}
-
-/**
- * Jitter strategy for retry delays to prevent thundering herd problem
- * @remarks
- * Jitter adds randomness to retry delays to spread out retry attempts when multiple operations fail simultaneously
- */
-export enum JitterStrategy {
-  /** No jitter - use exact calculated delay */
-  NONE = "NONE",
-  /** Full jitter - random delay between 0 and calculated delay */
-  FULL = "FULL",
-  /** Half jitter - random delay between 50% and 100% of calculated delay */
-  HALF = "HALF",
-}
-
-/**
- * Configuration options for step operations
- */
-export interface StepConfig<T> {
-  /** Strategy for retrying failed step executions */
-  retryStrategy?: (error: Error, attemptCount: number) => RetryDecision;
-  /** Execution semantics for the step (at-most-once or at-least-once per retry) */
-  semantics?: StepSemantics;
-  /** Serialization/deserialization configuration for step data */
-  serdes?: Serdes<T>;
-}
-
-/**
- * Configuration options for child context operations
- */
-export interface ChildConfig<T> {
-  /** Serialization/deserialization configuration for child context data */
-  serdes?: Serdes<T>;
-  /** Sub-type identifier for categorizing child contexts */
-  subType?: string;
-  /** Function to generate summaries for large results (used internally by map/parallel) */
-  summaryGenerator?: (result: T) => string;
-}
-
-/**
- * Configuration options for createCallback operations
- */
-export interface CreateCallbackConfig<T> {
-  /** Maximum time to wait for callback submission in seconds */
-  timeout?: number; // seconds
-  /** Heartbeat timeout in seconds to detect stalled callback operations */
-  heartbeatTimeout?: number; // seconds
-  /** Serialization/deserialization configuration for callback data */
-  serdes?: Serdes<T>;
-}
-
-/**
- * Configuration options for waitForCallback operations
- */
-export interface WaitForCallbackConfig<T> {
-  /** Maximum time to wait for callback in seconds */
-  timeout?: number; // seconds
-  /** Heartbeat timeout in seconds to detect stalled operations */
-  heartbeatTimeout?: number; // seconds
-  /** Strategy for retrying failed callback submissions */
-  retryStrategy?: (error: Error, attemptCount: number) => RetryDecision;
-  /** Serialization/deserialization configuration for callback data */
-  serdes?: Serdes<T>;
-}
-
-/**
- * Configuration options for invoke operations
- */
-export interface InvokeConfig<I, O> {
-  /** Serialization/deserialization configuration for input payload */
-  payloadSerdes?: Serdes<I>;
-  /** Serialization/deserialization configuration for result data */
-  resultSerdes?: Serdes<O>;
-}
-
-export type CreateCallbackResult<T> = [Promise<T>, string]; // [promise, callbackId]
-/**
- * Function that submits a callback ID to an external system
- * @param callbackId - Unique identifier for the callback that should be submitted to external system
- * @param context - Context for logging and other operations during callback submission
- * @returns Promise that resolves when the callback ID has been successfully submitted
- */
-export type WaitForCallbackSubmitterFunc = (
-  callbackId: string,
-  context: WaitForCallbackContext,
-) => Promise<void>;
-
-/**
- * Generic logger interface for custom logger implementations
- * Provides structured logging capabilities for durable execution contexts
- */
-export interface Logger {
-  /** Generic log method with configurable level */
-  log(level: string, message?: string, data?: unknown, error?: Error): void;
-  /** Log error messages with optional error object and additional data */
-  error(message?: string, error?: Error, data?: unknown): void;
-  /** Log warning messages with optional additional data */
-  warn(message?: string, data?: unknown): void;
-  /** Log informational messages with optional additional data */
-  info(message?: string, data?: unknown): void;
-  /** Log debug messages with optional additional data */
-  debug(message?: string, data?: unknown): void;
-}
-
-/**
- * Base interface for operation contexts.
- * Do not use directly - use specific context types like StepContext, WaitForConditionContext, etc.
- */
-export interface OperationContext {
-  logger: Logger;
-}
-
-export type StepContext = OperationContext;
-
-export type WaitForConditionContext = OperationContext;
-
-export type WaitForCallbackContext = OperationContext;
-
-/**
- * Function to be executed as a durable step
- * @param context - Context for logging and other operations during step execution
- * @returns Promise resolving to the step result
- */
-export type StepFunc<T> = (context: StepContext) => Promise<T>;
-/**
- * Function to be executed in a child context with isolated state
- * @param context - DurableContext with isolated step counter and state tracking
- * @returns Promise resolving to the child function result
- */
-export type ChildFunc<T> = (context: DurableContext) => Promise<T>;
-/**
- * Function to be executed for each item in a map operation
- * @param context - DurableContext for executing durable operations within the map
- * @param item - Current item being processed
- * @param index - Index of the current item in the array
- * @param array - The original array being mapped over
- * @returns Promise resolving to the transformed value
- */
-export type MapFunc<TInput, TOutput> = (
-  context: DurableContext,
-  item: TInput,
-  index: number,
-  array: TInput[],
-) => Promise<TOutput>;
-/**
- * Function to be executed as a branch in a parallel operation
- * @param context - DurableContext for executing durable operations within the branch
- * @returns Promise resolving to the branch result
- */
-export type ParallelFunc<T> = (context: DurableContext) => Promise<T>;
-
-/**
- * Named parallel branch with optional custom name
- */
-export interface NamedParallelBranch<T> {
-  name?: string;
-  func: ParallelFunc<T>;
-}
-
-/**
- * Function that checks and updates state for waitForCondition operations
- * @param state - Current state value
- * @param context - Context for logging and other operations during state checking
- * @returns Promise resolving to the updated state
- */
-export type WaitForConditionCheckFunc<T> = (
-  state: T,
-  context: WaitForConditionContext,
-) => Promise<T>;
-
-/**
- * Function that determines whether to continue waiting and how long to delay
- * @param state - Current state value
- * @param attempt - Current attempt number (starts at 1)
- * @returns Decision object indicating whether to continue and delay duration
- */
-export type WaitForConditionWaitStrategyFunc<T> = (
-  state: T,
-  attempt: number,
-) => WaitForConditionDecision;
-
-/**
- * Decision object for waitForCondition wait strategy
- */
-export type WaitForConditionDecision =
-  | { shouldContinue: true; delaySeconds: number }
-  | { shouldContinue: false };
-
-/**
- * Configuration options for waitForCondition operations
- */
-export interface WaitForConditionConfig<T> {
-  /** Strategy function that determines when to continue waiting and how long to delay */
-  waitStrategy: WaitForConditionWaitStrategyFunc<T>;
-  /** Initial state value to start the condition checking with */
-  initialState: T;
-  /** Serialization/deserialization configuration for state data */
-  serdes?: Serdes<T>;
-}
-
-/**
- * Configuration options for map operations
- */
-export interface MapConfig<TItem, TResult> {
-  /** Maximum number of concurrent executions (default: unlimited) */
-  maxConcurrency?: number;
-  /** Function to generate custom names for map items */
-  itemNamer?: (item: TItem, index: number) => string;
-  /** Serialization/deserialization configuration for parent context */
-  serdes?: Serdes<BatchResult<TResult>>;
-  /** Serialization/deserialization configuration for each item */
-  itemSerdes?: Serdes<TResult>;
-  /** Configuration for completion behavior */
-  completionConfig?: {
-    /** Minimum number of successful executions required */
-    minSuccessful?: number;
-    /** Maximum number of failures tolerated */
-    toleratedFailureCount?: number;
-    /** Maximum percentage of failures tolerated (0-100) */
-    toleratedFailurePercentage?: number;
-  };
-}
-
-/**
- * Configuration options for parallel operations
- */
-export interface ParallelConfig<TResult> {
-  /** Maximum number of concurrent executions (default: unlimited) */
-  maxConcurrency?: number;
-  /** Serialization/deserialization configuration for parent context */
-  serdes?: Serdes<BatchResult<TResult>>;
-  /** Serialization/deserialization configuration for each branch */
-  itemSerdes?: Serdes<TResult>;
-  /** Configuration for completion behavior */
-  completionConfig?: {
-    /** Minimum number of successful executions required */
-    minSuccessful?: number;
-    /** Maximum number of failures tolerated */
-    toleratedFailureCount?: number;
-    /** Maximum percentage of failures tolerated (0-100) */
-    toleratedFailurePercentage?: number;
-  };
-}
-
-/**
- * Represents an item to be executed with metadata for deterministic replay
- */
-export interface ConcurrentExecutionItem<T> {
-  /** Unique identifier for the item */
-  id: string;
-  /** The actual data/payload for the item */
-  data: T;
-  /** Index of the item in the original array */
-  index: number;
-  /** Optional custom name for the item */
-  name?: string;
-}
-
-/**
- * Executor function type for concurrent execution
- */
-export type ConcurrentExecutor<TItem, TResult> = (
-  item: ConcurrentExecutionItem<TItem>,
-  childContext: DurableContext,
-) => Promise<TResult>;
-
-/**
- * Represents a single item in a batch result
- */
-export interface BatchItem<R> {
-  /** The result value if the item succeeded */
-  result?: R;
-  /** The error if the item failed */
-  error?: Error;
-  /** Index of the item in the original array */
-  index: number;
-  /** Status of the item execution */
-  status: BatchItemStatus;
-}
-
-/**
- * Result of a batch operation (map, parallel, or concurrent execution)
- */
-export interface BatchResult<R> {
-  /** All items in the batch with their results/errors */
-  all: Array<BatchItem<R>>;
-
-  /** Returns only the items that succeeded */
-  succeeded(): Array<BatchItem<R> & { result: R }>;
-  /** Returns only the items that failed */
-  failed(): Array<BatchItem<R> & { error: Error }>;
-  /** Returns only the items that are still in progress */
-  started(): Array<BatchItem<R> & { status: BatchItemStatus.STARTED }>;
-
-  /** Overall status of the batch (SUCCEEDED if no failures, FAILED otherwise) */
-  status: BatchItemStatus.SUCCEEDED | BatchItemStatus.FAILED;
-  /** Reason why the batch completed */
-  completionReason:
-    | "ALL_COMPLETED"
-    | "MIN_SUCCESSFUL_REACHED"
-    | "FAILURE_TOLERANCE_EXCEEDED";
-  /** Whether any item in the batch failed */
-  hasFailure: boolean;
-
-  /** Throws the first error if any item failed */
-  throwIfError(): void;
-  /** Returns array of all successful results */
-  getResults(): Array<R>;
-  /** Returns array of all errors */
-  getErrors(): Array<Error>;
-
-  /** Number of successful items */
-  successCount: number;
-  /** Number of failed items */
-  failureCount: number;
-  /** Number of started but not completed items */
-  startedCount: number;
-  /** Total number of items */
-  totalCount: number;
-}
-
-/**
- * Configuration options for concurrent execution operations
- */
-export interface ConcurrencyConfig<TResult> {
-  /** Maximum number of concurrent executions (default: unlimited) */
-  maxConcurrency?: number;
-  /** Top-level operation subtype for tracking */
-  topLevelSubType?: string;
-  /** Iteration-level operation subtype for tracking */
-  iterationSubType?: string;
-  /** Function to generate summary from batch result */
-  summaryGenerator?: (result: BatchResult<TResult>) => string;
-  /** Serialization/deserialization configuration for parent context */
-  serdes?: Serdes<BatchResult<TResult>>;
-  /** Serialization/deserialization configuration for each item */
-  itemSerdes?: Serdes<TResult>;
-  /** Configuration for completion behavior */
-  completionConfig?: {
-    /** Minimum number of successful executions required */
-    minSuccessful?: number;
-    /** Maximum number of failures tolerated */
-    toleratedFailureCount?: number;
-    /** Maximum percentage of failures tolerated (0-100) */
-    toleratedFailurePercentage?: number;
-  };
-}
+export * from "./core";
+export * from "./logger";
+export * from "./step";
+export * from "./child-context";
+export * from "./callback";
+export * from "./invoke";
+export * from "./wait-condition";
+export * from "./batch";
+export * from "./durable-context";
