+++ packages/aws-durable-execution-sdk-js/src/types/invoke.ts
@@ -1,0 +1,11 @@
+import { Serdes } from "../utils/serdes/serdes";
+
+/**
+ * Configuration options for invoke operations
+ */
+export interface InvokeConfig<I, O> {
+  /** Serialization/deserialization configuration for input payload */
+  payloadSerdes?: Serdes<I>;
+  /** Serialization/deserialization configuration for result data */
+  resultSerdes?: Serdes<O>;
+}
